Prevent errors calling #event_bus or #repository
Those methods are only available within the Sinatra application, nowhere
else. For instance, the first version of the TransactionalEventBus had a
mistake where `event_bus` was being used, instead of `@event_bus`. This
caused a stack overflow. This new version prevents this problem by
scoping the #event_bus and #repository methods to the Sinatra
application only.