Add welcome email to registration flow
In order to simplify the code, I abstracted the worker state to a single
table for all workers. The state is now a single row per worker, in the
same table. At the same time, the code to get and set the worker state
was identical, except for the different names. This too was abstracted.